Move db related logic to edge_worker model
