+++ providers/edge3/tests/unit/edge3/cli/test_edge_command.py
@@ -433,14 +433,13 @@ class TestEdgeWorkerCli:
 
     @pytest.mark.db_test
     def test_list_edge_workers(self, mock_edgeworker: EdgeWorkerModel):
-        mock_session = MagicMock()
         args = self.parser.parse_args(["edge", "list-workers", "--output", "json"])
         with contextlib.redirect_stdout(StringIO()) as temp_stdout:
             with patch(
-                "airflow.providers.edge3.cli.edge_command._fetch_edge_hosts_from_db",
+                "airflow.providers.edge3.models.edge_worker.get_registered_edge_hosts",
                 return_value=[mock_edgeworker],
             ):
-                edge_command.list_edge_workers(args, session=mock_session)
+                edge_command.list_edge_workers(args)
                 out = temp_stdout.getvalue()
                 edge_workers = json.loads(out)
         for key in [
